feat: add data-story/core version
